create controllers that wrap entities methods
