+++ backend/src/controllers/OrphanagesController.ts
@@ -1,0 +1,35 @@
+import { Request, Response } from 'express';
+import { getRepository } from 'typeorm';
+import Orphanage from '../models/Orphanage';
+
+const OrphanagesController = {
+  async create(request: Request, response: Response)  {
+    const {
+      name,
+      latitude,
+      longitude,
+      about,
+      instructions,
+      opening_hours,
+      open_on_weekends
+    } = request.body;
+  
+    const orphanagesRepository = getRepository(Orphanage);
+  
+    const orphanage = orphanagesRepository.create({
+      name,
+      latitude,
+      longitude,
+      about,
+      instructions,
+      opening_hours,
+      open_on_weekends
+    });
+  
+    await orphanagesRepository.save(orphanage);
+  
+    return response.status(201).json(orphanage);
+  }
+};
+
+export default OrphanagesController;
