+++ backend/src/routes.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import OrphanagesController from './controllers/OrphanagesController';
+
+const routes = Router();
+
+routes.post('/orphanages', OrphanagesController.create);
+
+export default routes;
